+++ src/Expressions/Operators/Operator.php
@@ -7,6 +7,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 namespace WikibaseSolutions\CypherDSL\Expressions\Operators;
 
 use WikibaseSolutions\CypherDSL\QueryConvertible;
@@ -38,7 +39,7 @@ abstract class Operator implements QueryConvertible
     /**
      * Whether to insert parentheses around the given expression, given the precedence of this operator.
      */
-    final protected function shouldInsertParentheses(AnyType $expression): bool
+    protected function shouldInsertParentheses(AnyType $expression): bool
     {
         if (!$expression instanceof self) {
             return false;