+++ tests/unit/Expressions/Operators/GreaterThanOrEqualTest.php
@@ -7,6 +7,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 namespace WikibaseSolutions\CypherDSL\Tests\Unit\Expressions\Operators;
 
 use PHPUnit\Framework\TestCase;
@@ -27,7 +28,7 @@ final class GreaterThanOrEqualTest extends TestCase
 
         $greaterThanOrEqual = new GreaterThanOrEqual($greaterThanOrEqual, $greaterThanOrEqual);
 
-        $this->assertSame("(10 >= 15) >= (10 >= 15)", $greaterThanOrEqual->toQuery());
+        $this->assertSame("10 >= 15 >= 10 >= 15", $greaterThanOrEqual->toQuery());
     }
 
     public function testInstanceOfBooleanType(): void