+++ tests/unit/Expressions/Operators/MultiplicationTest.php
@@ -7,6 +7,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 namespace WikibaseSolutions\CypherDSL\Tests\Unit\Expressions\Operators;
 
 use PHPUnit\Framework\TestCase;
@@ -34,14 +35,14 @@ final class MultiplicationTest extends TestCase
 
     public function testInstanceOfFloatType(): void
     {
-        $multiplication = new Multiplication(new Float_(10.0), new Float_(15.0), false);
+        $multiplication = new Multiplication(new Float_(10.0), new Float_(15.0));
 
         $this->assertInstanceOf(FloatType::class, $multiplication);
     }
 
     public function testInstanceOfIntegerType(): void
     {
-        $multiplication = new Multiplication(new Integer(10), new Integer(15), false);
+        $multiplication = new Multiplication(new Integer(10), new Integer(15));
 
         $this->assertInstanceOf(IntegerType::class, $multiplication);
     }