+++ tests/unit/QueryTest.php
@@ -7,6 +7,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 namespace WikibaseSolutions\CypherDSL\Tests\Unit;
 
 use Iterator;
@@ -303,7 +304,7 @@ final class QueryTest extends TestCase
 
         $pathMock = new Path([$nodeMock, (new Node)->withVariable('b')], [new Relationship(Direction::RIGHT)]);
         $numeralMock = new Integer(12);
-        $booleanMock = new GreaterThan($variableMock, new Variable('b'), false);
+        $booleanMock = new GreaterThan($variableMock, new Variable('b'));
         $propertyMock = new Property($variableMock, 'b');
 
         $query = new Query();